+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # CHANGELOG
+* `3.4.7` Release - [3.4.7](#3.4.7) &nbsp; - &nbsp; 2022-08-09
 * `3.4.6` Release - [3.4.6](#3.4.6) &nbsp; - &nbsp; 2022-05-23
 * `3.4.5` Release - [3.4.5](#3.4.5) &nbsp; - &nbsp; 2022-04-13
 * `3.4.4` Release - [3.4.4](#3.4.4) &nbsp; - &nbsp; 2022-04-04
@@ -23,6 +24,15 @@
 * `3.1.8` Release - [3.1.8](#3.1.8) &nbsp; - &nbsp; 2020-12-11
 * `3.1.7` Release - [3.1.7](#3.1.7) &nbsp; - &nbsp; 2020-12-04
 * `3.1.6` Release - [3.1.6](#3.1.6) &nbsp; - &nbsp; 2020-11-13
+
+# <a name="3.4.7"></a> 3.4.7
+#### Features
+`
+DevEnigmaRiverAndroidCore
+DevEnigmaRiverAndroidDownload
+`
+* EMP-18520- download in progress asset
+* EMP-18524- improved error message
 
 # <a name="3.4.6"></a> 3.4.6
 #### Features